requirements.txt
Flask
pymongo
RPi.GPIO
datetime
json
bson
flask
uwsgi